requirements.txt
python-dotenv
pandas
numpy
torch==2.4.0.dev20240414
torchaudio==2.2.0.dev20240414
torchvision==0.19.0.dev20240414
